test: separate scripts out into file
